Better error message for function-not-found if there is a function having the same name
PostgreSQL always raises the function/procedure not found error if it finds ambiguity in the function name or arguments. The error message can be misleading if a function or procedure already exists with the same name, but a different set of parameters.

This commit adds report_proc_not_found_error_hook hook to re-analyse the procedure/function call and generate a more meaningful error message depending upon procedure/function arguments to mimic T-SQL behaviour. If we are still unable to locate a function/procedure in the hook, then Babelfish will let PostgreSQL handle the error message.

Author: Rishabh Tanwar <ritanwar@amazon.com>
Signed-off-by: Sharu Goel <goelshar@amazon.com>