Fix the issue where the coalesce function returns incorrect typmod for sys.binary type
PostgreSQL's coalesce function internally coerces each argument to a common type. 
For example, if type is domain, then it's get coerced to it's base type. When the 
argument is of type sys.binary (which is a domain type), it is coerced to a sys.bbf_binary 
(the base type for sys.binary). During coercion from domain to base type PostgreSQL 
takes the default typmod for type instead of getting it from source domain type. This is 
expected as PostgreSQL does not allow domains to have a typmod.

Since T-SQL allows a typmod for domains, we now transfer typmod information from 
source type to target type with the coalesce function when using T-SQL dialect.

Task: BABEL-1087
Author: Rishabh Tanwar <ritanwar@amazon.com>
Signed-off-by: Xiaohui Fanhe <hexiaohu@amazon.com>